Modern builds PokemonSubstruct0 differently than agbcc, trying fix

+++ include/pokemon.h
@@ -19,8 +19,8 @@ struct PokemonSubstruct0
     /*0x08*/ u8 ppBonuses;
     /*0x09*/ u8 friendship;
     /*0x0A*/ u16 pokeball:5; //31 balls
-             u8 hiddenNature:5;  // 25 natures
-             u8 free_sub0_b:6;
+             u16 hiddenNature:5;  // 25 natures
+             u16 free_sub0_b:6;
 }; /* size = 12 */
